+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/interfaces/core/CanEqual.kt
@@ -1,4 +1,4 @@
-package com.windea.breezeframework.core.interfaces
+package com.windea.breezeframework.core.interfaces.core
 
 interface CanEqual {
 	override fun equals(other: Any?): Boolean