+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/interfaces/core/CanToString.kt
@@ -1,4 +1,4 @@
-package com.windea.breezeframework.core.interfaces
+package com.windea.breezeframework.core.interfaces.core
 
 interface CanToString {
 	override fun toString(): String