+++ breeze-dsl/src/main/kotlin/com/windea/breezeframework/dsl/markup/MarkdownDsl.kt
@@ -3,7 +3,6 @@
 package com.windea.breezeframework.dsl.markup
 
 import com.windea.breezeframework.core.extensions.*
-import com.windea.breezeframework.core.interfaces.*
 import com.windea.breezeframework.dsl.*
 import com.windea.breezeframework.dsl.markup.MarkdownConfig.addPrefixHeadingMarkers
 import com.windea.breezeframework.dsl.markup.MarkdownConfig.emptyColumnLength
@@ -292,7 +291,7 @@ class MarkdownWikiLink @PublishedApi internal constructor(
 @MarkdownDsl
 class MarkdownTextBlock @PublishedApi internal constructor(
 	val text: String
-) : MarkdownDslTopElement, WrapContent {
+) : MarkdownDslTopElement, CanWrap {
 	override var wrapContent: Boolean = true
 	
 	override fun toString(): String {
@@ -307,7 +306,7 @@ class MarkdownTextBlock @PublishedApi internal constructor(
 sealed class MarkdownHeading(
 	val headingLevel: Int,
 	val text: String
-) : MarkdownDslTopElement, WithMarkdownAttributes, WrapContent {
+) : MarkdownDslTopElement, WithMarkdownAttributes, CanWrap {
 	override var attributes: MarkdownAttributes? = null
 	override var wrapContent: Boolean = true
 }
@@ -418,7 +417,7 @@ class MarkdownList @PublishedApi internal constructor(
 sealed class MarkdownListNode(
 	protected val prefixMarkers: String,
 	val text: String
-) : MarkdownDslElement, WrapContent {
+) : MarkdownDslElement, CanWrap {
 	val nodes: MutableList<MarkdownListNode> = mutableListOf()
 	
 	override var wrapContent: Boolean = true
@@ -459,7 +458,7 @@ class MarkdownTaskListNode @PublishedApi internal constructor(
 @MarkdownDslExtendedFeature
 class MarkdownDefinition @PublishedApi internal constructor(
 	val title: String
-) : MarkdownDslTopElement, WrapContent {
+) : MarkdownDslTopElement, CanWrap {
 	val nodes: MutableList<MarkdownDefinitionNode> = mutableListOf()
 	
 	override var wrapContent: Boolean = true
@@ -480,7 +479,7 @@ class MarkdownDefinition @PublishedApi internal constructor(
 @MarkdownDslExtendedFeature
 class MarkdownDefinitionNode @PublishedApi internal constructor(
 	val text: String
-) : MarkdownDslElement, WrapContent {
+) : MarkdownDslElement, CanWrap {
 	override var wrapContent: Boolean = true
 	
 	override fun toString(): String {
@@ -513,6 +512,10 @@ class MarkdownTable @PublishedApi internal constructor() : MarkdownDslTopElement
 		val rowsSnippet = rows.joinToStringOrEmpty("\n")
 		return "$headerRowSnippet\n$delimitersSnippet\n$rowsSnippet"
 	}
+	
+	enum class Alignment(val textPair: Pair<String, String>) {
+		None("-" to "-"), Left(":" to "-"), Center(":" to ":"), Right("-" to ":")
+	}
 }
 
 /**Markdown表格头部。*/
@@ -544,7 +547,7 @@ class MarkdownTableHeader @PublishedApi internal constructor() : MarkdownDslElem
 	override fun String.unaryPlus() = column(this)
 	
 	@MarkdownDsl
-	inline infix fun MarkdownTableColumn.align(alignment: MarkdownTableAlignment) =
+	inline infix fun MarkdownTableColumn.align(alignment: MarkdownTable.Alignment) =
 		this.also { it.alignment = alignment }
 }
 
@@ -573,7 +576,7 @@ open class MarkdownTableRow @PublishedApi internal constructor() : MarkdownDslEl
 class MarkdownTableColumn @PublishedApi internal constructor(
 	val text: String = emptyColumnText
 ) : MarkdownDslElement {
-	var alignment: MarkdownTableAlignment = MarkdownTableAlignment.None //only for columns in table header
+	var alignment: MarkdownTable.Alignment = MarkdownTable.Alignment.None //only for columns in table header
 	
 	override fun toString(): String {
 		return text
@@ -664,9 +667,9 @@ class MarkdownMultilineMath @PublishedApi internal constructor(
 @MarkdownDsl
 @MarkdownDslExtendedFeature
 class MarkdownAdmonition @PublishedApi internal constructor(
-	val qualifier: MarkdownAdmonitionQualifier,
+	val qualifier: Qualifier,
 	val title: String = "",
-	val type: MarkdownAdmonitionType = MarkdownAdmonitionType.Normal
+	val type: Type = Type.Normal
 ) : MarkdownDslTopElement, MarkdownDslEntry {
 	override val content: MutableList<MarkdownDslTopElement> = mutableListOf()
 	
@@ -676,6 +679,25 @@ class MarkdownAdmonition @PublishedApi internal constructor(
 		val titleSnippet = title.wrapQuote(quote)
 		val contentSnippet = toContentString().prependIndent(indent)
 		return "${type.text} ${qualifier.text} $titleSnippet\n$contentSnippet"
+	}
+	
+	enum class Type(val text: String) {
+		Normal("!!!"), Collapsed("???"), Opened("!!!+")
+	}
+	
+	enum class Qualifier(val style: String, val text: String) {
+		Abstract("abstract", "abstract"), Summary("abstract", "summary"), Tldr("abstract", "tldr"),
+		Bug("bug", "bug"),
+		Danger("danger", "danger"), Error("danger", "error"),
+		Example("example", "example"), Snippet("example", "snippet"),
+		Fail("fail", "fail"), Failure("fail", "failure"), Missing("fail", "missing"),
+		Question("fag", "question"), Help("fag", "help"), Fag("fag", "fag"),
+		Info("info", "info"), Todo("info", "todo"),
+		Note("note", "note"), SeeAlso("note", "seealso"),
+		Quote("quote", "quote"), Cite("quote", "cite"),
+		Success("success", "success"), Check("success", "check"), Done("success", "done"),
+		Tip("tip", "tip"), Hint("tip", "hint"), Important("tip", "important"),
+		Warning("warning", "warning"), Caution("warning", "caution"), Attention("warning", "attention")
 	}
 }
 
@@ -693,7 +715,7 @@ class MarkdownFrontMatter @PublishedApi internal constructor(
 /**Markdown目录。只能位于文档顶部。用于生成当前文档的目录。*/
 @MarkdownDsl
 @MarkdownDslExtendedFeature
-class MarkdownToc @PublishedApi internal constructor() : MarkdownDslElement, GenerateContent {
+class MarkdownToc @PublishedApi internal constructor() : MarkdownDslElement, CanGenerate {
 	override var generateContent: Boolean = false
 	
 	override fun toGeneratedString(): String {
@@ -711,7 +733,7 @@ class MarkdownToc @PublishedApi internal constructor() : MarkdownDslElement, Gen
 @MarkdownDslExtendedFeature
 class MarkdownImport @PublishedApi internal constructor(
 	val url: String
-) : MarkdownDslTopElement, GenerateContent, WithMarkdownAttributes {
+) : MarkdownDslTopElement, CanGenerate, WithMarkdownAttributes {
 	//DONE extended classes and properties
 	override var attributes: MarkdownAttributes? = null
 	override var generateContent: Boolean = false
@@ -733,7 +755,7 @@ class MarkdownImport @PublishedApi internal constructor(
 @MarkdownDslExtendedFeature
 class MarkdownMacros @PublishedApi internal constructor(
 	val name: String
-) : MarkdownDslTopElement, GenerateContent {
+) : MarkdownDslTopElement, CanGenerate {
 	override var generateContent: Boolean = false
 	
 	override fun toGeneratedString(): String {
@@ -764,7 +786,9 @@ class MarkdownMacrosSnippet @PublishedApi internal constructor(
 @MarkdownDsl
 sealed class MarkdownReference(
 	val reference: String
-) : MarkdownDslElement, CanEqual
+) : MarkdownDslElement, WithName {
+	override val _name: String get() = reference
+}
 
 /**Markdown脚注的引用。*/
 @MarkdownDsl
@@ -863,39 +887,6 @@ class MarkdownPropertyAttribute(
 	override fun toString(): String {
 		return "$name=${value.wrapQuote(quote)}"
 	}
-}
-
-//REGION enumerations and constants
-
-/**Markdown表格的对齐方式。*/
-@MarkdownDslExtendedFeature
-enum class MarkdownTableAlignment(val textPair: Pair<String, String>) {
-	None("-" to "-"), Left(":" to "-"), Center(":" to ":"), Right("-" to ":")
-}
-
-/**Markdown警告框的类型。*/
-@MarkdownDsl
-@MarkdownDslExtendedFeature
-enum class MarkdownAdmonitionType(val text: String) {
-	Normal("!!!"), Collapsed("???"), Opened("!!!+")
-}
-
-/**Markdown警告框的限定符。*/
-@MarkdownDsl
-@MarkdownDslExtendedFeature
-enum class MarkdownAdmonitionQualifier(val style: String, val text: String) {
-	Abstract("abstract", "abstract"), Summary("abstract", "summary"), Tldr("abstract", "tldr"),
-	Bug("bug", "bug"),
-	Danger("danger", "danger"), Error("danger", "error"),
-	Example("example", "example"), Snippet("example", "snippet"),
-	Fail("fail", "fail"), Failure("fail", "failure"), Missing("fail", "missing"),
-	Question("fag", "question"), Help("fag", "help"), Fag("fag", "fag"),
-	Info("info", "info"), Todo("info", "todo"),
-	Note("note", "note"), SeeAlso("note", "seealso"),
-	Quote("quote", "quote"), Cite("quote", "cite"),
-	Success("success", "success"), Check("success", "check"), Done("success", "done"),
-	Tip("tip", "tip"), Hint("tip", "hint"), Important("tip", "important"),
-	Warning("warning", "warning"), Caution("warning", "caution"), Attention("warning", "attention")
 }
 
 //REGION build extensions
@@ -1107,8 +1098,8 @@ inline fun MarkdownDslEntry.multilineMath(lazyText: () -> String) =
 
 @MarkdownDsl
 @MarkdownDslExtendedFeature
-inline fun MarkdownDslEntry.admonition(qualifier: MarkdownAdmonitionQualifier, title: String = "",
-	type: MarkdownAdmonitionType = MarkdownAdmonitionType.Normal, block: MarkdownAdmonition.() -> Unit) =
+inline fun MarkdownDslEntry.admonition(qualifier: MarkdownAdmonition.Qualifier, title: String = "",
+	type: MarkdownAdmonition.Type = MarkdownAdmonition.Type.Normal, block: MarkdownAdmonition.() -> Unit) =
 	MarkdownAdmonition(qualifier, title, type).also { it.block() }.also { content += it }
 
 @MarkdownDsl