+++ CHANGELOG.md
@@ -157,7 +157,10 @@
 * [X] breeze-dsl 修复不足之处。
 - [X] breeze-dsl 优化Dsl构建方法。
 * [X] breeze-dsl 完成`FlowChartDsl`，完善其他的Dsl。
-* [ ] breeze-dsl 实现`SequenceDiagramDsl`。（过于复杂，搁置）
+* [X] breeze-dsl 实现`SequenceDiagramDsl`，将枚举尽可能地移到对应的类内部。
+* [ ] breeze-dsl 实现`MermaidStateDiagram`。
+* [ ] breeze-dsl 优化：可省略的元素声明，其对应的元素要实现`CanOmitDeclaration`，且可进行过滤。
+* [ ] breeze-dsl 优化：对于表示转换的元素，可以通过`"a"(...) fromTo "b"`的语法构建。
 * [ ] breeze-dsl 实现`YamlDsl`。（不推荐使用，但是作为规范）
 
 ***